Revert "Record short array. Easier to deal with."
This reverts commit 0a647463b5df7dd2c83e4174d6317244c95abb99.

+++ app/src/main/java/com/climbtheworld/app/walkietalkie/IntercomBackgroundService.java
@@ -101,7 +101,7 @@ public class IntercomBackgroundService extends Service implements IClientEventLi
 			}
 
 			@Override
-			public void onAudio(final short[] frame, int numberOfReadBytes, double energy, double rms) {
+			public void onAudio(final byte[] frame, int numberOfReadBytes, double energy, double rms) {
 
 			}
