+++ app/src/main/java/com/climbtheworld/app/walkietalkie/audiotools/IVoiceDetector.java
@@ -1,5 +1,5 @@
 package com.climbtheworld.app.walkietalkie.audiotools;
 
 public interface IVoiceDetector {
-	boolean onAudio(short[] frame, int numberOfReadBytes, double energy);
+	boolean onAudio(byte[] frame, int numberOfReadBytes, double energy);
 }
