+++ app/src/main/java/com/climbtheworld/app/walkietalkie/states/PushToTalkState.java
@@ -51,7 +51,7 @@ public class PushToTalkState extends InterconState implements IInterconState, IR
 	}
 
 	@Override
-	public void onAudio(short[] frame, int numberOfReadBytes, double energy, double rms) {
+	public void onAudio(byte[] frame, int numberOfReadBytes, double energy, double rms) {
 		updateEnergy(energy);
 	}
